Update GameEnvironment.java
So if we have no food, i made it so the player waste his time looking for food when there is no food in his inventory. Also, the food is now removed from the inventory was eaten by the pet. and it now cost action points to feed the pet.